use float instead of decimal

+++ shoppingcart/abc.py
@@ -1,6 +1,5 @@
 import abc
 import typing
-from decimal import Decimal
 
 class ShoppingCart(abc.ABC):
 
@@ -16,5 +15,5 @@ class ShoppingCart(abc.ABC):
 
 class ProductPriceDataSource(abc.ABC):
     @abc.abstractmethod
-    def get_product_price(self, product_code: str) -> Decimal:
+    def get_product_price(self, product_code: str) -> float:
         pass
