+++ shoppingcart/product_prices.py
@@ -1,7 +1,6 @@
 import json
 import csv
 from . import abc
-from decimal import Decimal
 
 
 class DefaultProductPrices(abc.ProductPriceDataSource):
@@ -12,7 +11,7 @@ class DefaultProductPrices(abc.ProductPriceDataSource):
             "kiwi" : 3.0
         }
     
-    def get_product_price(self, product_code: str) -> Decimal:
+    def get_product_price(self, product_code: str) -> float:
         return self._prices.get(product_code, 0.0)
 
 
@@ -21,8 +20,8 @@ class JSONProductPrices(abc.ProductPriceDataSource):
         with open(filename, 'r') as file:
             self._prices = json.load(file)
 
-    def get_product_price(self, product_code: str) -> Decimal:
-        return Decimal(self._prices.get(product_code, 0.0))
+    def get_product_price(self, product_code: str) -> float:
+        return self._prices.get(product_code, 0.0)
 
 
 class CSVProductPrices(abc.ProductPriceDataSource):
@@ -31,8 +30,8 @@ class CSVProductPrices(abc.ProductPriceDataSource):
         with open(filename, 'r') as file:
             reader = csv.DictReader(file)
             for row in reader:
-                self._prices[row['product_code']] = Decimal(row['price'])
+                self._prices[row['product_code']] = float(row['price'])
 
 
-    def get_product_price(self, product_code: str) -> Decimal:
-        return Decimal(self._prices.get(product_code, 0.0))
+    def get_product_price(self, product_code: str) -> float:
+        return self._prices.get(product_code, 0.0)
